requirements.txt
ansible==2.8.2
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyYAML==5.1.1
ruamel.yaml==0.15.99
six==1.12.0
Werkzeug==0.15.4
WTForms==2.2.1
